Exception when operating with arrays
Pilate therefore said to him, "Are you a king then?" Jesus answered, "You say that I am a king. For this reason I have been born, and for this reason I have come into the world, that I should testify to the truth. Everyone who is of the truth listens to my voice." (John 18:37)